create home directory and move components
