+++ react-app-mui/src/pages/home/Home.tsx
@@ -1,4 +1,4 @@
-import Navbar from "../components/navbar/Navbar";
+import Navbar from "../../components/navbar/Navbar";
 import {Outlet} from "react-router-dom";
 import HomeContainer from "./HomeContainer.styled";
