+++ react-app-mui/src/pages/home/HomeContainer.styled.ts
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import {S_SIZE} from "../helpers/constants";
+import {S_SIZE} from "../../helpers/constants";
 
 
 const HomeContainer = styled.div`
